requirements.txt
# Core dependencies for Talishar data downloader
requests>=2.31.0
python-dotenv>=1.0.0

# Optional dependencies for data analysis
pandas>=2.0.0